refactor(ext-http): extract client event helpers to client_events.rs (2000-line gate)
